Update RARPC.py
I noticed that there is also a need to check again for rpc availability in this part because sometimes, when the discord app got closed right after it check its connectivity and then it updates presence, it will be stuck in a loop of trying to update it without checking again if the pipe can be connected or not. I appreciate your concerns about error catching.